# openapi.yaml
openapi: 3.1.0
info:
  title: SEMD API
  version: 1.0.0
paths:
...
  /auth/register:
    post:
      tags:
      - auth
      summary: User Registration
      description: Register a new user account with MEMBER role. Public endpoint.
      operationId: register_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
  /auth/create:
    post:
      tags:
      - auth
      summary: Create Admin User
      description: Create a new ADMIN user. Requires SUPER_ADMIN role.
      operationId: create_user_auth_create_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUserResponse'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
  /auth/login:
    post:
      tags:
      - auth
      summary: User Login
      description: Authenticate user with username and password. Returns token pair
        or pre-auth token if 2FA is enabled.
      operationId: login_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/TokenPairResponse'
                - $ref: '#/components/schemas/PreAuthResponse'
                title: Response Login Auth Login Post
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
  /auth/login/2fa:
    post:
      tags:
      - auth
      summary: Complete 2FA Login
      description: Complete authentication with 2FA verification code after receiving
        pre-auth token.
      operationId: login_2fa_auth_login_2fa_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TwoFAVerifyRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPairResponse'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
  /auth/login/provider:
    post:
      tags:
      - auth
      summary: OAuth Provider Login
      description: Authenticate using OAuth provider (GitHub or Google) token.
      operationId: login_provider_auth_login_provider_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginProviderRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/TokenPairResponse'
                - $ref: '#/components/schemas/PreAuthResponse'
                title: Response Login Provider Auth Login Provider Post
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
  /auth/refresh:
    post:
      tags:
      - auth
      summary: Refresh Access Token
      description: Exchange refresh token for a new token pair.
      operationId: refresh_auth_refresh_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPairResponse'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
  /auth/logout:
    post:
      tags:
      - auth
      summary: User Logout
      description: Revoke refresh token and logout user.
      operationId: logout_auth_logout_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponseModel'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
  /auth/2fa/setup:
    post:
      tags:
      - auth
      summary: Setup 2FA
      description: Generate TOTP secret and QR code for 2FA setup. Requires authentication.
      operationId: setup_2fa_auth_2fa_setup_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TwoFASetupResponse'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
  /auth/2fa/enable:
    post:
      tags:
      - auth
      summary: Enable 2FA
      description: Enable 2FA for the authenticated user by verifying OTP code. Requires
        authentication.
      operationId: enable_2fa_auth_2fa_enable_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TwoFAEnableRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponseModel'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
  /auth/oauth/device:
    post:
      tags:
      - auth
      summary: Initiate OAuth Device Flow
      description: Start OAuth device authorization flow for GitHub. Returns device
        code and user code.
      operationId: oauth_device_initiate_auth_oauth_device_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthDeviceCodeRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthDeviceCodeResponse'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
  /auth/oauth/device/poll:
    post:
      tags:
      - auth
      summary: Poll OAuth Device Flow
      description: Poll for OAuth device flow completion. Returns token pair when
        authorized or pending status.
      operationId: oauth_device_poll_auth_oauth_device_poll_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthDevicePollRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/TokenPairResponse'
                - $ref: '#/components/schemas/PreAuthResponse'
                - $ref: '#/components/schemas/BaseResponseModel'
                title: Response Oauth Device Poll Auth Oauth Device Poll Post
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
  /auth/oauth/authorize:
    post:
      tags:
      - auth
      summary: Get OAuth Authorization URL
      description: Generate OAuth authorization URL for GitHub or Google login.
      operationId: oauth_authorize_auth_oauth_authorize_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthAuthorizationRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthAuthorizationResponse'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
  /auth/callback/{provider}:
    get:
      tags:
      - auth
      summary: OAuth Callback
      description: Handle OAuth callback from provider. Exchanges authorization code
        for tokens.
      operationId: oauth_callback_auth_callback__provider__get
      parameters:
      - name: provider
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/OAuthProviderType'
      - name: code
        in: query
        required: true
        schema:
          type: string
          title: Code
      - name: state
        in: query
        required: true
        schema:
          type: string
          title: State
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/TokenPairResponse'
                - $ref: '#/components/schemas/PreAuthResponse'
                title: Response Oauth Callback Auth Callback  Provider  Get
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
  /auth/me:
    get:
      tags:
      - auth
      summary: Get My Profile
      description: Get current user's profile information. Requires authentication.
      operationId: get_me_auth_me_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserModel'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
    put:
      tags:
      - auth
      summary: Update My Profile
      description: Update current user's profile information. Requires authentication.
      operationId: update_me_auth_me_put
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserModel'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
  /auth/me/reset-password:
    post:
      tags:
      - auth
      summary: Reset My Password
      description: Reset current user's password. Requires current password verification.
      operationId: reset_password_auth_me_reset_password_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponseModel'
        '404':
          description: Not found
        '401':
          description: Unauthorized
        '422':
          description: Validation error
...
components:
  schemas:
...
    AuthLoginProviderRequest:
      properties:
        provider:
          $ref: '#/components/schemas/OAuthProviderType'
          title: provider
          description: Authentication provider (e.g., google, github)
        token:
          type: string
          title: token
          description: Authentication token from provider
      type: object
      required:
      - provider
      - token
      title: AuthLoginProviderRequest
    AuthLoginRequest:
      properties:
        username:
          type: string
          title: username
          description: Username for login
        password:
          type: string
          title: password
          description: Password for login
      type: object
      required:
      - username
      - password
      title: AuthLoginRequest
    BaseResponseModel:
      properties:
        status:
          type: integer
          title: status
          description: Response status code
          examples:
          - 200
          - 404
          - 500
        message:
          type: string
          title: message
          description: Response message
          examples:
          - Success
          - Not found
          - Internal server error
      type: object
      required:
      - status
      - message
      title: BaseResponseModel
    CreateUserRequest:
      properties:
        username:
          type: string
          maxLength: 32
          minLength: 3
          title: username
          description: Username for the new user
        email:
          type: string
          format: email
          title: email
          description: Email address
        full_name:
          type: string
          maxLength: 512
          title: full_name
          description: Full name of the user
        password:
          type: string
          minLength: 8
          title: password
          description: Password for the account
        role:
          $ref: '#/components/schemas/RoleType'
          title: role
          description: Role for the new user (ADMIN only)
          default: ADMIN
      type: object
      required:
      - username
      - email
      - full_name
      - password
      title: CreateUserRequest
    CreateUserResponse:
      properties:
        status:
          type: integer
          title: Status
          default: 201
        message:
          type: string
          title: Message
          default: User created successfully
        user_id:
          type: integer
          title: user_id
          description: ID of the newly created user
      type: object
      required:
      - user_id
      title: CreateUserResponse
...
    OAuthAuthorizationRequest:
      properties:
        provider:
          $ref: '#/components/schemas/OAuthProviderType'
          title: provider
          description: OAuth provider (github or google)
      type: object
      required:
      - provider
      title: OAuthAuthorizationRequest
    OAuthAuthorizationResponse:
      properties:
        status:
          type: integer
          title: Status
          default: 200
        message:
          type: string
          title: Message
          default: Authorization URL generated
        authorization_url:
          type: string
          title: authorization_url
          description: URL to redirect user for authorization
        state:
          type: string
          title: state
          description: State parameter for CSRF protection
      type: object
      required:
      - authorization_url
      - state
      title: OAuthAuthorizationResponse
    OAuthDeviceCodeRequest:
      properties:
        provider:
          $ref: '#/components/schemas/OAuthProviderType'
          title: provider
          description: OAuth provider (github or google)
      type: object
      required:
      - provider
      title: OAuthDeviceCodeRequest
    OAuthDeviceCodeResponse:
      properties:
        status:
          type: integer
          title: Status
          default: 200
        message:
          type: string
          title: Message
          default: Device code generated
        device_code:
          type: string
          title: device_code
          description: Device verification code
        user_code:
          type: string
          title: user_code
          description: User verification code to enter
        verification_uri:
          type: string
          title: verification_uri
          description: URL where user enters the code
        expires_in:
          type: integer
          title: expires_in
          description: Seconds until code expires
        interval:
          type: integer
          title: interval
          description: Minimum seconds between polling requests
      type: object
      required:
      - device_code
      - user_code
      - verification_uri
      - expires_in
      - interval
      title: OAuthDeviceCodeResponse
    OAuthDevicePollRequest:
      properties:
        provider:
          $ref: '#/components/schemas/OAuthProviderType'
          title: provider
          description: OAuth provider
        device_code:
          type: string
          title: device_code
          description: Device code to poll
      type: object
      required:
      - provider
      - device_code
      title: OAuthDevicePollRequest
    OAuthProviderType:
      type: string
      enum:
      - github
      - google
      title: OAuthProviderType
...
    PreAuthResponse:
      properties:
        status:
          type: integer
          title: Status
          default: 200
        message:
          type: string
          title: Message
          default: 2FA verification required
        pre_auth_token:
          type: string
          title: pre_auth_token
          description: Pre-authentication token for 2FA
        requires_2fa:
          type: boolean
          title: requires_2fa
          description: Indicates 2FA is required
          default: true
      type: object
      required:
      - pre_auth_token
      title: PreAuthResponse
...
    RefreshTokenRequest:
      properties:
        refresh_token:
          type: string
          title: refresh_token
          description: Refresh token to exchange for new token pair
      type: object
      required:
      - refresh_token
      title: RefreshTokenRequest
    RegisterRequest:
      properties:
        username:
          type: string
          maxLength: 32
          minLength: 3
          title: username
          description: Username for registration
        email:
          type: string
          format: email
          title: email
          description: Email address
        full_name:
          type: string
          maxLength: 512
          title: full_name
          description: Full name of the user
        password:
          type: string
          minLength: 8
          title: password
          description: Password for the account
      type: object
      required:
      - username
      - email
      - full_name
      - password
      title: RegisterRequest
    RegisterResponse:
      properties:
        status:
          type: integer
          title: Status
          default: 201
        message:
          type: string
          title: Message
          default: User registered successfully
        user_id:
          type: integer
          title: user_id
          description: ID of the newly created user
      type: object
      required:
      - user_id
      title: RegisterResponse
...
    RoleType:
      type: string
      enum:
      - GUEST
      - MEMBER
      - ADMIN
      - SUPER_ADMIN
      title: RoleType
...
    TokenPairResponse:
      properties:
        status:
          type: integer
          title: Status
          default: 200
        message:
          type: string
          title: Message
          default: Authentication successful
        access_token:
          type: string
          title: access_token
          description: JWT access token
        refresh_token:
          type: string
          title: refresh_token
          description: JWT refresh token
        token_type:
          type: string
          title: token_type
          description: Token type
          default: bearer
      type: object
      required:
      - access_token
      - refresh_token
      title: TokenPairResponse
...
    TwoFAEnableRequest:
      properties:
        secret:
          type: string
          title: secret
          description: TOTP secret key to enable
        otp_code:
          type: string
          title: otp_code
          description: One-time password code to verify
      type: object
      required:
      - secret
      - otp_code
      title: TwoFAEnableRequest
    TwoFASetupResponse:
      properties:
        status:
          type: integer
          title: Status
          default: 200
        message:
          type: string
          title: Message
          default: 2FA setup initiated
        secret:
          type: string
          title: secret
          description: TOTP secret key
        qr_uri:
          type: string
          title: qr_uri
          description: QR code URI for authenticator apps
      type: object
      required:
      - secret
      - qr_uri
      title: TwoFASetupResponse
    TwoFAVerifyRequest:
      properties:
        pre_auth_token:
          type: string
          title: pre_auth_token
          description: Pre-authentication token
        otp_code:
          type: string
          title: otp_code
          description: One-time password code
      type: object
      required:
      - pre_auth_token
      - otp_code
      title: TwoFAVerifyRequest
...
    UserUpdateRequest:
      properties:
        username:
          anyOf:
          - type: string
            maxLength: 32
            minLength: 3
          - type: 'null'
          title: Username
          description: Username
        email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Email
          description: Email address
        full_name:
          anyOf:
          - type: string
            maxLength: 512
            minLength: 1
          - type: 'null'
          title: Full Name
          description: Full name
        birthday:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Birthday
          description: Birthday
        profile_img_uri:
          anyOf:
          - type: string
          - type: 'null'
          title: Profile Img Uri
          description: Profile image URI
      type: object
      title: UserUpdateRequest
...